+++ src/neuroconv/tools/__init__.py
@@ -1,1 +1,2 @@
 from .importing import get_package
+from .path_expansion import LocalPathExpander
